+++ backend/src/handlers/index.ts
@@ -80,40 +80,6 @@ export const login = async (req: Request, res: Response) => {
 
 // el jsonwebtoken se debe enviar en los headers en el bearer
 export const getUser = async(req: Request, res: Response) => {
-    const bearer = req.headers.authorization
-    //console.log('bearer', bearer)
-     if(!bearer) {
-         const error = new Error('No autorizado')
-         res.status(401).json({error: error.message})
-         return
-     }
-     console.log('entro')
-     const [, token] = bearer.split(' ')
-     if(!token) {
-        const error = new Error('No autorizado')
-        res.status(401).json({error: error.message})
-        return
-    }
-    // verificar eltoken 
-    try {
-        const result = jwt.verify(token, process.env.JWT_SECRET)
-        if (typeof result ==='object' && result.id){
-            console.log('result.id', result.id)
-            //const user = await User.findById(result.id).select('name handle email')
-            const user = await User.findById(result.id).select('-password')
-            console.log('user', user)
-            if (!user) {
-                const error = new Error('Usuario no existe')
-                res.status(404).json({error: error.message})
-            }
-            res.status(200).json({message: user})
-            return 
-        }
-    } catch (error) {
-        console.log('error', error)
-        res.status(500).json({error: 'token no valido'})
-        return 
-    }
-     //console.log(token)
-     
+     console.log('desde get use', req.user)
+     res.send(req.user) 
 }
