+++ backend/src/router.ts
@@ -3,6 +3,7 @@ import { Router } from "express";
 import { createAccount, getUser, login } from "./handlers";
 import { body } from "express-validator";
 import { handleInputErrors } from "./midleware/validation";
+import { authenticate } from "./midleware/auth";
 
 const router = Router()
  
@@ -33,5 +34,6 @@ router.post('/auth/login',
     handleInputErrors,
     login)
 
-router.get('/user', getUser)
+router.get('/user', authenticate, getUser)
+
 export default router;
