+++ frontend/src/api/DevTreeaPI.ts
@@ -1,0 +1,13 @@
+import { isAxiosError } from "axios";
+import api from "../config/axios";
+
+export async function getUser() {
+    try {
+        const { data } = await api.post(`/auth/login`, formData);
+        localStorage.setItem("AUTH_TOKEN", data);
+      } catch (error) {
+        if (isAxiosError(error) && error.response) {
+          console.log("error", error.response.data.error);
+        }
+      }
+}
